+++ internal/provider/provider.go
@@ -263,6 +263,7 @@ func (p *datahubProvider) Resources(_ context.Context) []func() resource.Resourc
 	return []func() resource.Resource{
 		NewConnectionResource,
 		NewCorpGroupResource,
+		NewCorpGroupMemberResource,
 		NewIngestionSourceResource,
 		NewSecretResource,
 		NewRemoteExecutorPoolResource,